Issue #73 fix: rename post -> write post
