--- app/src/main/java/com/gdsc/sogongsogong/ui/post/PostAdapter.kt
@@ -1,4 +1,0 @@
-package com.gdsc.sogongsogong.ui.post
-
-class PostAdapter {
-}
